requirements.txt
aiogram>=3.0.0,<4.0.0
aiosqlite>=0.17.0
aiohttp>=3.8.1
python-dotenv>=0.21.0
